requirements.txt
fastapi
uvicorn
sentence-transformers
xgboost
scikit-learn
python-docx
PyPDF2
numpy
pandas
